populate_db/requirements.txt
pandas==1.3.5
python-dotenv==0.21.0
psycopg2-binary==2.9.3
requests==2.28